requirements.txt
matplotlib~=3.2.2
numpy~=1.20.1
sklearn~=0.0
scikit-learn~=0.23.2
joblib~=1.0.1
pyyaml~=5.4.1
visualdl~=2.2.0
librosa~=0.9.1
tqdm~=4.59.0